requirements.txt
Django>=5.2
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.0
psycopg2-binary>=2.9
redis>=5.0
django-redis>=5.4
django-rq>=2.10
whitenoise>=6.6
gunicorn>=21.2
Pillow>=10.0
python-dotenv>=1.0
